+++ src/services/files/file.hooks.js
@@ -1,26 +1,11 @@
 const commonHooks = require('feathers-hooks-common');
 const defaultHooks = require('./file_default.hook.js');
 const authenticateHook = require('../../authentication/authenticationHook');
+const { skipInternal, getCurrentUserData } = require('../../authentication/permissionHelper.hooks.js');
 const { unifySlashes } = require('../../hooks/unifySlashes');
 
 const errors = require('@feathersjs/errors');
 
-const getCurrentUserData = hook => {
-  const userModel = hook.app.get('mongooseClient').model('users');
-  return new Promise((resolve, reject) => {
-    userModel.findById(hook.params.user._id, function(err, user) {
-      if (err) {
-        reject(new errors.GeneralError(err));
-      }
-      if (!user) {
-        reject(new errors.NotFound('User not found'));
-      }
-      hook.params.user.role = user.role;
-      hook.params.user.providerId = user.providerId;
-      return resolve(hook);
-    });
-  });
-};
 
 const restrictResourceToCurrentProvider = async hook => {
   if (hook.params.user.role !== 'superhero') {
@@ -77,13 +62,6 @@ const checkPermissionsAfterAuthentication = hook => {
         return restrictToPublicResources(hook);
       });
   });
-};
-
-const skipInternal = method => hook => {
-  if (typeof hook.params.provider === 'undefined') {
-    return hook;
-  }
-  return method(hook);
 };
 
 const distributionHooks = {
